requirements.txt
requests
pillow
openai
python-dotenv